% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAndCleanData.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans NOAA data frame with earthquake related data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{a data frame to clean}
}
\value{
Cleaned data frame
}
\description{
It's expected that the data frame is produced by the \code{\link{file_read}}
function. The cleaning is performed in the following way:
\enumerate{
 \item Creates the \code{DATE} column by uniting the year, month, day
   and converting it to the Date class
 \item Converts the \code{LATITUDE} and \code{LONGITUDE} columns to numeric class
 \item Cleans the \code{LOCATION_NAME} column by stripping out the country name
   (including the colon) and converts names to title case (as opposed to all caps)}
}
\examples{
library(dplyr)
file_read() \%>\% eq_clean_data()

}
