% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Prepare the annotation of an earthquake}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A row of data frame data containing earthquake related data.}
}
\value{
An HTML label, which can be used as the annotation text in the leaflet map
}
\description{
The function creates an HTML label, which can be used as the annotation text in the leaflet map.
The function combines the following information about an earthquake:
\itemize{
 \item Location
 \item Magnitude
 \item Total number of deaths}
If an earthquake is missing values for any of these items, it is skipped for that element
of the tag.
}
\examples{
file_read() \%>\%
  eq_clean_data() \%>\%
  NOAAC \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2014) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}
