% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAndCleanData.R
\name{eq_get_date}
\alias{eq_get_date}
\title{Date Conversion from Numeric Representation}
\usage{
eq_get_date(Y, M = 1, D = 1)
}
\arguments{
\item{Y, }{numerical values to specify a year.}

\item{M, }{numerical values to specify a month}

\item{D, }{numerical values to specify a day}
}
\value{
An object of class "\code{\link{POSIXct}}".
}
\description{
In contrast to \code{\link{ISOdate}} it can also define BC dates (at least those
of them that comply with R). Only the additional day of leap years (February 29)
cannot be defined correctly as R introduces extra 0-th year, which actually didn't
exist. That is the only reason why BC leap years in R are shifted on 1-year away
from real ones.
}
\examples{
 eq_get_date(-100, 12, 31);
 eq_get_date(2000);
}
