% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAndCleanData.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans the location data}
\usage{
eq_location_clean(location)
}
\arguments{
\item{location}{input location to clean.}
}
\value{
Cleaned location which contains the places affected by earthquake converted
  to title case.
}
\description{
It's supposed to be used to clean the data of the LOCATION_NAME column. It does
it in the following way:
\itemize{
 \item stripes out the country name (including the colon)
 \item converts names to title case (as opposed to all caps)}
}
\examples{
 eq_location_clean("TURKEY:  IZMIR, EFES"); #Izmir, Efes
}
