% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{eq_map}
\alias{eq_map}
\title{Visualize earthquakes on a map}
\usage{
eq_map(eq_data, annot_col)
}
\arguments{
\item{eq_data}{A data frame containing earthquake related data.}

\item{annot_col}{The column name which should be used for annotation purpose}
}
\description{
The function takes an argument \code{eq_data} containing the filtered data frame
with earthquakes to visualize. The function maps the epicenters
(\code{LATITUDE}/\code{LONGITUDE}) and annotates each point with in pop up window
containing annotation data stored in a column of the data frame.
The user is able to choose which column to use for the annotation in
the pop-up with a function argument named \code{annot_col}.
Each earthquake is shown with a circle whose radius is proportional
to the earthquake's magnitude.
}
\examples{
file_read() \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "CHINA" & lubridate::year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")
}
