% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAndCleanData.R
\name{file_read}
\alias{file_read}
\title{Read earthquake related data from a file}
\usage{
file_read(filename = "inst\\\\extdata\\\\results.txt")
}
\arguments{
\item{filename}{A name of the file to read. By default read the data
of an embedded file (downloaded from \href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{NOAA database}).}
}
\description{
The function is supposed to be used to read an embedded file which contains
earthquake related data
(downloaded from \href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{NOAA database}).
It also can be used in case you need to read a file downloaded manually.
}
\examples{
\dontrun{file_read()}

}
