% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimelineGeoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{A time line of earthquakes}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show.legend = NA, na.rm = FALSE,
  inherit.aes = TRUE, xmin = NA, xmax = NA, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{xmin}{Min date for additional filtering}

\item{xmax}{Max date for additional filtering}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
This geom plots a time line of earthquakes.
ranging from \code{xmin} to \code{xmax} dates with a point for each earthquake.
Optional aesthetics include color, size, and alpha (for transparency).
The \code{x} aesthetic is a date and an optional \code{y} aesthetic is a
factor indicating some stratification in which case multiple time lines
will be plotted for each level of the factor (e.g. country).
}
\section{Aesthetics}{


\code{geom_timeline} understands the following aesthetics
(required aesthetics are in bold):
\itemize{
 \item \strong{x}
 \item y
 \item xmin
 \item xmax
 \item alpha
 \item colour
 \item fill
 \item size
 \item shape}
}

\examples{
#without stratification
file_read() \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY \%in\% c("USA", "CHINA"),
                between(lubridate::year(DATE), 2010, 2011)) \%>\%
  ggplot(aes(x = DATE,
             color = DEATHS,
             size = EQ_PRIMARY)) +
  geom_timeline(alpha = 0.4) +
  theme_timeline() +
  labs(size = "Richter scale value", colour = "# deaths")

#with stratification and additional filtering by time
file_read() \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY \%in\% c("USA", "CHINA", "CHILE"),
                between(lubridate::year(DATE), 2010, 2011)) \%>\%
  ggplot(aes(x = DATE,
             y = COUNTRY,
             color = DEATHS,
             size = EQ_PRIMARY)) +
  geom_timeline(xmin = eq_get_date(2010, 1, 1),
                xmax = eq_get_date(2010, 12, 31),
  theme_timeline() +
  labs(size = "Richter scale value", colour = "# deaths")
}
