% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimelineGeoms.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Annotations to the earthquake data}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, n_max = NA, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{n_max}{A number of largest (by magnitude) earthquakes to be annotated.
If omitted, each earthquake is annotated}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
This geom adds a vertical line to each data point with a text annotation
(e.g. the location of the earthquake) attached to each line. There is an
option to subset to \code{n_max} number of earthquakes, where we take the
\code{n_max} largest (by magnitude) earthquakes.
Aesthetics are \code{x}, which is the date of the earthquake and \code{label}
which takes the column name from which annotations will be obtained.
If \code{n_max} is omitted, each earthquake is annotated
}
\section{Aesthetics}{


\code{geom_timeline_label} understands the following aesthetics
(required aesthetics are in bold):
\itemize{
 \item \strong{x}
 \item \strong{label}
 \item n_max}
}

\examples{
#without stratification with 2 greatest earthquakes are annotated
file_read() \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY \%in\% c("USA", "CHINA"),
                between(lubridate::year(DATE), 2010, 2011)) \%>\%
  ggplot(aes(x = DATE,
             color = DEATHS,
             size = EQ_PRIMARY)) +
  geom_timeline(alpha = 0.4) +
  theme_timeline() +
  labs(size = "Richter scale value", colour = "# deaths") +
  geom_timeline_label(aes(label = LOCATION_NAME), n_max = 2)
}
