% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimelineGeoms.R
\name{theme_timeline}
\alias{theme_timeline}
\title{This theme is used to visualize earthquakes}
\usage{
theme_timeline()
}
\description{
The theme is built on top \code{\link{theme_bw}} in order to follow the style
of the Coursera's assignment
}
\examples{
ggplot(mtcars) + geom_point(aes(x = wt, y = mpg)) + theme_timeline()

}
